Chinh sua giao dien quan ly

+++ src/com/raven/form/QuanLyThuPhi/Form_LichSuDanhSachDienNuoc.java
@@ -60,7 +60,7 @@ public class Form_LichSuDanhSachDienNuoc extends javax.swing.JPanel {
             }
         ) {
             Class[] types = new Class [] {
-                java.lang.Integer.class, java.lang.Integer.class, java.lang.String.class
+                java.lang.Integer.class, java.lang.Long.class, java.lang.String.class
             };
             boolean[] canEdit = new boolean [] {
                 false, false, false
